Rename `LatLonGridIterator` to `RegularGridIterator` for future enhancement
