+++ src/grid/helpers.rs
@@ -4,6 +4,33 @@ use proj::Proj;
 #[allow(unused_imports)]
 use crate::{GribError, GridPointIndexIterator};
 
+/// An iterator over latitudes and longitudes of grid points of a regular grid.
+#[derive(Clone)]
+pub struct RegularGridIterator {
+    lat: Vec<f32>,
+    lon: Vec<f32>,
+    ij: GridPointIndexIterator,
+}
+
+impl RegularGridIterator {
+    pub(crate) fn new(lat: Vec<f32>, lon: Vec<f32>, ij: GridPointIndexIterator) -> Self {
+        Self { lat, lon, ij }
+    }
+}
+
+impl Iterator for RegularGridIterator {
+    type Item = (f32, f32);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let (i, j) = self.ij.next()?;
+        Some((self.lat[j], self.lon[i]))
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        self.ij.size_hint()
+    }
+}
+
 #[cfg(feature = "gridpoints-proj")]
 pub(crate) fn latlons_from_projection_definition_and_first_point(
     proj_def: &str,
@@ -102,5 +129,88 @@ pub(crate) mod test_helpers {
     pub(crate) fn assert_coord_almost_eq((x1, y1): (f32, f32), (x2, y2): (f32, f32), delta: f32) {
         assert_almost_eq!(x1, x2, delta);
         assert_almost_eq!(y1, y2, delta);
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::{super::ScanningMode, *};
+
+    macro_rules! test_lat_lon_grid_iter {
+        ($(($name:ident, $scanning_mode:expr, $expected:expr),)*) => ($(
+            #[test]
+            fn $name() {
+                let lat = (0..3).into_iter().map(|i| i as f32).collect::<Vec<_>>();
+                let lon = (10..12).into_iter().map(|i| i as f32).collect::<Vec<_>>();
+                let scanning_mode = ScanningMode($scanning_mode);
+                let ij= GridPointIndexIterator::new(lon.len(), lat.len(), scanning_mode);
+                let actual = RegularGridIterator::new(lat, lon, ij).collect::<Vec<_>>();
+                assert_eq!(actual, $expected);
+            }
+        )*);
+    }
+
+    test_lat_lon_grid_iter! {
+        (
+            lat_lon_grid_iter_with_scanning_mode_0b00000000,
+            0b00000000,
+            vec![
+                (0., 10.),
+                (0., 11.),
+                (1., 10.),
+                (1., 11.),
+                (2., 10.),
+                (2., 11.),
+            ]
+        ),
+        (
+            lat_lon_grid_iter_with_scanning_mode_0b00100000,
+            0b00100000,
+            vec![
+                (0., 10.),
+                (1., 10.),
+                (2., 10.),
+                (0., 11.),
+                (1., 11.),
+                (2., 11.),
+            ]
+        ),
+        (
+            lat_lon_grid_iter_with_scanning_mode_0b00010000,
+            0b00010000,
+            vec![
+                (0., 10.),
+                (0., 11.),
+                (1., 11.),
+                (1., 10.),
+                (2., 10.),
+                (2., 11.),
+            ]
+        ),
+        (
+            lat_lon_grid_iter_with_scanning_mode_0b00110000,
+            0b00110000,
+            vec![
+                (0., 10.),
+                (1., 10.),
+                (2., 10.),
+                (2., 11.),
+                (1., 11.),
+                (0., 11.),
+            ]
+        ),
+    }
+
+    #[test]
+    fn lat_lon_grid_iterator_size_hint() {
+        let lat = (0..3).map(|i| i as f32).collect::<Vec<_>>();
+        let lon = (10..12).map(|i| i as f32).collect::<Vec<_>>();
+        let scanning_mode = ScanningMode(0b00000000);
+        let ij = GridPointIndexIterator::new(lon.len(), lat.len(), scanning_mode);
+        let mut iter = RegularGridIterator::new(lat, lon, ij);
+
+        assert_eq!(iter.size_hint(), (6, Some(6)));
+        let _ = iter.next();
+        assert_eq!(iter.size_hint(), (5, Some(5)));
     }
 }
